chore: move generate user object tests to its own file
